requirements.txt
Django==1.8
argparse==1.2.1
django-factory-boy==1.0.0
mock==2.0.0
wsgiref==0.1.2
